created the parquet/encryption/[external] folder to put external encryptor related stuff. Begin moving things around.
