+++ cpp/src/parquet/CMakeLists.txt
@@ -241,8 +241,8 @@ if(PARQUET_REQUIRE_ENCRYPTION)
   list(APPEND PARQUET_STATIC_LINK_LIBS ${ARROW_OPENSSL_LIBS})
   set(PARQUET_SRCS ${PARQUET_SRCS} encryption/encryption_internal.cc
                    encryption/openssl_internal.cc
-                   encryption/dll_encryptor.cc
-                   encryption/loadable_encryptor_utils.cc)
+                   encryption/external/dll_encryptor.cc
+                   encryption/external/loadable_encryptor_utils.cc)
   # Encryption key management
   set(PARQUET_SRCS
       ${PARQUET_SRCS}
